Split tests into module for each struct
This is the structure that I initially wanted, an individual module for
each struct, or set of functions which implements all the tests for
that collection of related items. This puts the tests right next to the
implementation so both can be edited and seen at once which is really
nice.