Test cell::to_cartesian function and fix bugs
This adds testing to the cell.to_cartesian function, taking an Isometry
and using the cell parameters to convert the relative coordinates to
real space coordinates. This tests some simple cases, fixing some bugs
in the calculation of the transformation, namely taking the sin and cos
of the cell angle.

Additionally this adds a default cell, being a cell with sides of 1, 1,
and angle of PI/2.