Take command line arguments to configure simulation
This uses the clap library for the parsing of command line arguments,
allowing the configuration of many options at runtime. This makes it a
whole lot easier to use and test different configurations.

The current options available are;
- wallpaper group -> which of the wallpaper groups in the enum of all
wallpaper groups to use. This checks it is one of the allowed groups in
the parsing of the argument.
- num-sides -> the number of sides the shape should have. This allows
for quickly testing the different polygons.
- steps -> the number of monte carlo steps to run.