Parallel implementation of multiple simulations
This parallelises finding the packing fraction using rayon. A new state
is created for each thread, for which the best packing is then found. I
have implemented the Ord trait on the PackedState to then be able to
find the best configuration.

Using rayon required that I switch to Arg and Mutex to hold the shared
values, which does appear to have a small impact on performance on small
shapes.