Refactor Line impls for clarity in Intersects
This is primarily a refactoring of the Line struct to provide clarity on
the implementation of the Intersects trait. This is a trait to allow a
range of definitions of a shape to behave in much the same way. Taking a
list of something that can check for intersections and comparing all
against each other. This isn't the fastest method in all cases,
(particularly lines) however it is simple and adaptable.