Use Shape trait to generalise implementation details
I am not just interested in the implementation of a specific definition
of a shape. I would also like to look at shapes defined with circles in
both 2 and 3 dimensions, and not prevent the investigation of further
types of shapes.

Using a trait to define the necessary implementation details means that
in theory, anyone can extend upon my definitions of a shape in their own
library, although here it means I can define a collection of shapes, all
which have different implementation details, yet all have the same
interface.

Here a Shape is a list of things which can be checked for an
intersection. In this case I have a LineShape consisting of Line
segments and a MolecularShape consisting of Atoms, defined as points
with radii. This removes the original RadialShape struct, although there
is a constructor for the LineShape from a radial definition. In moving
to the LineShape there is a ten fold speedup in the calculation of
intersections, so it is definitely staying.