Improve calculation of MolShape area to be useful
The calculation of area in the MolecularShape was the super naive area
of all shapes. This adds support for the calculation of overlap between
two circles, however it will overcompensate where there is overlap of
more than two circles. While not a perfect algorithm this is still
useful since the shapes I am currently using only have overlaps of two
components. The handling of the general case can be done at a later
point in time.