Bugfix: Fix overlapping shapes
There were occasions where shapes would overlap, giving packings of
greater than 1, which is impossible. There were a couple of bugs here.

Firstly the incorrect check for the range of the packing fraction, which
has a check for being in the correct range. Here I was using an OR
instead of an AND for the conditional, meaning the values over 1 were
not an error.

The second bug was only checking the first neighbour shell for
intersections. When the cell tilts there are intersections which aren't
picked up by comparing the central particle with the nearest neighbour
shell, so using the second shell is required. These still show up in the
images of the cells as intersections between the first neighbour shell.
Expanding the search to the second neighbour shell is fewer comparisons
than comparing all the first neighbour shells with each other.