Handle equal cell lengths within Cell
This provides a collection of abstractions, which allow for the handling
of Cells with equal lengths entirely within the cell module, i.e. there
is no multiple ownership of a SharedValue within a cell. In doing so
this provides abstractions; x(), y(), and angle() which get the relevant
floating point values for these quantities, no more .get_value().

This also includes documentation of a number of methods within the cell
module.