Use my own implementation of a symmetry transform
This moves the implementation of the IsometryMatrix2 to the Symmetry
transform struct for a few reasons. One of the main reasons is that the
IsometryMatrix2 does not intentionally support mirror operations, while
I was able to implement them using the from_matrix_unchecked, I would
prefer not to rely on undefined behaviour. Secondly it was really
difficult to work with them in practice for my use case. Adding the
translations, transforming from relative to real coordinates was messy.
Having my own implementation means I can make those parts of the
implementation simpler, namely adding support for the Add trait.

The updated implementation is still backed by the nalgebra vectors and
matrices, just now a little easer to work with.