maint: remove basis from PackedState instance
The basis is a collection of values for manipulating the state, it isn't
required when the state it not being mutated, i.e. the Basis values can
live within a function. This removes the basis values from a PackedState
instance, generating them when required from a method on the
PackedState.