enh: Replace Arc/Mutex shared value with raw pointers
The SharedValue struct is now using the dereferencing of raw pointers
for it's values. This is an unsafe method, however this is the exact
reason for having unsafe bocks.

In making this change there are a number changes to how items are passed
around in functions. Most notably a PackedState is 'consumed' by the
monte carlo process which makes a lot more sense since the mutated state
was useless.