maint: more stringent requirements for the Shape trait
This is a little complicated. For the shape trait, I require that the
component of the shape is able to undertake multiplication by both
reference and value. Unfortunately this is not entirely simple, with the
requirement having to be passed up the stack of calling values. This is
because there is no code path where this conditional doesn't hold, so
all the calling functions also need to have it.

Unfortunately there is currently no way to properly express this, so I
have to go with the messy version.