enh: Benchmarking the creation of a ShapeInstance
The creation of a ShapeInstance is one of the components in the checking
of an intersection for a shape so as such I would like to optimise it as
much as possible.

This is intended to provide information on what elements of the
comparison are taking up all the time.