refactor: Split lib.rs into smaller files
The lib.rs contained too much code which wasn't particularly related.
This breaks the code into a packing and site module which are now much
smaller. The additional benefit of this change is that that lib module
is now focused on the API of the module, which of the functions and
structs should be exported publicly in the base module. The idea is to
make it simple to use, with a flattening out of the structure. This
means that there isn't the requirement to remember where different
elements are located.