feat: Auto generate changelog from commit messages
Using the git-chglog tool I am now able to generate a changelog from the
git commit messages. This does require that the commit adhere to a
specific format, the conventional changelog style, which involves
starting the commit with the type of change followed by a semicolon,
i.e. one of
- fix:
- feat:
- chore:
- docs:
- style:
- refactor:
- perf:
- test: