refactor: Separate 2D and 3D into separate modules
This organises the 2D and the 3D components such that there is
separation and it makes sense. The handling to algorithms is different
for each dimension so it makes sense to treat them individually.

Where there is functionality that makes sense to share it can be
implemented as a trait.