refactor: Consistency with implementation of Intersection
The intersection trait is one that is not only applicable to the base
types but also the more complex shapes. This makes the Intersection
trait a requirement of the Shape trait and hence makes the
implementation specific to each instance of the shape now there is no
longer a default implementation.