refactor: PackedState handles intersections
This is the first step towards a more general approach to the
optimisation. This uses the PackedState to check for an intersection in
returning the score, with a valid score being the result and an
intersection being an error condition. This makes the approach much more
general, finding a score which is to be optimised.