refactor: Create State trait for the crystal state
The State trait is designed to be a general interface for the current
state of a simulation. Having a score, and the ability to get the basis,
that is all the degrees of freedom which can be changed. There is
additionally a total shapes value which allows for the appropriate
scaling of the temperature parameter.