refactor: Make area a function of intersect trait
A feature of having an intersection is that the area is the property
used for scoring. This makes sure there is an implementation of area for
any instance of the intersect trait, rather than making it a property of
a Shape, which could have any number of methods of scoring.

This starts to pave the way for Shapes constructed using potentials.