feat: Implement a Shape using the LJ potential
This is a shape using the Lennard Jones potential, where the score is
no longer based on the area, but instead the potential energy. There are
some issues here like the potential energy being long range and more
shells decreasing the energy, however I do hope this isn't too much of a
problem.

Regardless this is an initial implementation of the shape including a
new trait for handling shapes comprised of potentials.