feat: Incorporate potential energy further
This updates the potential energy minimisations so that the interface
for the shapes and the State is the same, independent on the type of
interaction required. This includes a number of fixes for the traits and
requirements of the states, this is really helping to solidify the parts
of the interface which are required and which are not necessary.

This has resulted in the removal of the interface trait from the Shape
class, instead requiring each State to specify additional requirements
separately.