fix: Scale score of potential for improved results
This gets better results from the search space by scaling the results of
the potential. When working with the packing it expects a score in the
range (0, 1) which is not the same as the default for the potential.
This scales the score of the potential to roughly reside in this range.