feat: Create simple output of positions
When dealing with other programs, the simplest method of defining the
structure is not the serialisation, which requires implementing a lot of
understanding about crystal structure theory. Instead the most useful
representation is the shape of the crystal cell and the positions and
orientations of the particles within the cell.