fix: Optimisation always returns valid state (#24)
This makes sure that the optimisation always returns a valid state, with
the condition that the user has to initially enter a valid state. This
is checked when the argument is first passed, exiting if the state is
invalid. Additionally there is something wrong if the state is invalid
at the end of the configuration so this is also checked.