refactor: Readability of arguments for Transform::new
This improves the readability of specifying the arguments of the
Transform2 class. This takes an angular rotation in radians as the first
argument, while the translation as a tuple is the second argument.

This makes the initialisation simple, and additionally there will be
fewer issues with the order or the arguments since the structure of the
types makes it explicit.