refactor: Simplify optimisation function
This is a drastic refactoring of the optimisation function, breaking it
up into a number of smaller parts each responsible for separate things.

The accept score function is now used to check for the acceptance of the
score, returning a Option, with some for an accepted score and None for
a rejected score. This leaves the optimise function to focus of the
effects of acceptance or rejection.