refactor: Convert SharedValue to using UnsafeCell
This refactoring makes a lot of changes to the way that a SharedValue is
constructed using a more idiomatic form of interior mutability and
making some of the assumptions more explicit.

The UnsafeCell is the fundamental construct of interior mutability, with
a range of wrappers providing different guarantees. I have provided a
very simple wrapper which currently provides no guarantees, whilst
implementing both the Send and Sync traits. The idea is that there will
be no issues based on the current usage of the crate.

Additionally the use of UnsafeCell provides more guarantees, like being
explicit about what has interior mutability, along with the appropriate
lifetime checks. These are all absent in the previous implementation,
which uses raw pointers and this implementation adds no overhead.