refactor: Use Option over Result for score return
This changes the return value of the score functions from the Result
type to an Option type. This is a breaking change.

The reason for this change is one of messaging. The Option type
indicates that None is a value which is an expected part of the
operation of the program, while the Result indicates that some kind of
error has occurred and so needs to be handled.

An None (or previously Err) score is one which is encountered in the
standard operation of the program, when hard disc particles overlap. For
this reason, the Option type is far more suitable.

BREAKING CHANGE
- score functions return an Option instead of a Result type.