refactor: Change internal rep of cell to ratio/length
This changes the representation of the cell from the length of each
side, to a length and the ratio for each side. This provides an
interface which is more aligned with the restrictions of the different
unit cells.

This additionally makes it possible to limit the ratio of the lengths
such that the a value of the unit cell is larger than the b value.