ref: Convert Basis from trait to enum
The Basis is an internal method that describes the actions that can be
performed on a range of values. The advantage of a trait is that you can
add new values outside of this crate, however that is something that
will never be done. So to allow for different bases to be used, I have
converted it to an enum, which is sized and can be added to a vectors.

There are a number of changes here including moving around where the
values can be modified.